mcp_client/requirements.txt
azure-identity>=1.23.1
mcp>=1.12.2
openai>=1.97.1
python-dotenv>=1.1.1
sseclient>=0.0.27